refact: rename feature modules to :feature-*
