+++ app/build.gradle
@@ -38,7 +38,7 @@ dependencies {
     implementation project(':logger')
     implementation project(':preferences')
     implementation project(':ui')
-    implementation project(':ui-history')
-    implementation project(':ui-today')
+    implementation project(':feature-history')
+    implementation project(':feature-today')
     implementation libraries.material
 }
