+++ app/src/main/java/com/jeanbarrossilva/h2o/H2OApplication.kt
@@ -2,8 +2,8 @@ package com.jeanbarrossilva.h2o
 
 import android.app.Application
 import com.jeanbarrossilva.h2o.injection.boundaryModule
-import com.jeanbarrossilva.h2o.ui.history.historyModule
-import com.jeanbarrossilva.h2o.ui.today.todayModule
+import com.jeanbarrossilva.h2o.feature.history.historyModule
+import com.jeanbarrossilva.h2o.feature.today.todayModule
 import org.koin.core.context.startKoin
 
 class H2OApplication: Application() {
