+++ app/src/main/java/com/jeanbarrossilva/h2o/injection/BoundaryModule.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.h2o.injection
 
 import com.jeanbarrossilva.h2o.boundary.DefaultTodayBoundary
-import com.jeanbarrossilva.h2o.ui.today.TodayBoundary
+import com.jeanbarrossilva.h2o.feature.today.TodayBoundary
 import org.koin.dsl.module
 
 val boundaryModule = module {
