+++ feature-history/src/main/res/drawable/ic_drink_water.png
