+++ feature-today/src/main/java/com/jeanbarrossilva/h2o/feature/today/TodayActivity.kt
@@ -1,0 +1,6 @@
+package com.jeanbarrossilva.h2o.feature.today
+
+import com.jeanbarrossilva.h2o.feature.today.R
+import com.jeanbarrossilva.h2o.ui.template.navigation.NavigationActivity
+
+class TodayActivity: NavigationActivity(R.navigation.today)
