+++ feature-today/src/main/java/com/jeanbarrossilva/h2o/feature/today/TodayBoundary.kt
@@ -1,0 +1,9 @@
+package com.jeanbarrossilva.h2o.feature.today
+
+import androidx.fragment.app.FragmentActivity
+
+interface TodayBoundary {
+    fun navigateToIntakeLogger()
+
+    fun navigateToHistory(activity: FragmentActivity?)
+}
