+++ feature-today/src/main/java/com/jeanbarrossilva/h2o/feature/today/TodayModule.kt
@@ -1,0 +1,10 @@
+package com.jeanbarrossilva.h2o.feature.today
+
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val todayModule = module {
+    viewModel {
+        TodayViewModel(preferenceManager = get(), logger = get())
+    }
+}
