+++ feature-today/src/main/res/values/strings.xml
