+++ settings.gradle
@@ -4,6 +4,6 @@ include ':logger'
 include ':calculator'
 include ':preferences'
 include ':ui'
-include ':ui-today'
+include ':feature-today'
 include ':model'
-include ':ui-history'
+include ':feature-history'
