--- ui-today/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest
-    package="com.jeanbarrossilva.h2o.ui.today"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-    <application>
-        <activity
-            android:name="com.jeanbarrossilva.h2o.ui.today.TodayActivity"
-            android:exported="false" />
-    </application>
-</manifest>
