--- ui-today/src/main/java/com/jeanbarrossilva/h2o/ui/today/TodayActivity.kt
@@ -1,5 +1,0 @@
-package com.jeanbarrossilva.h2o.ui.today
-
-import com.jeanbarrossilva.h2o.ui.template.navigation.NavigationActivity
-
-class TodayActivity: NavigationActivity(R.navigation.today)
