--- ui-today/src/main/java/com/jeanbarrossilva/h2o/ui/today/TodayBoundary.kt
@@ -1,9 +1,0 @@
-package com.jeanbarrossilva.h2o.ui.today
-
-import androidx.fragment.app.FragmentActivity
-
-interface TodayBoundary {
-    fun navigateToIntakeLogger()
-
-    fun navigateToHistory(activity: FragmentActivity?)
-}
