--- ui-today/src/main/res/navigation/today.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/today"
-    app:startDestination="@id/today_fragment">
-    <fragment
-        android:id="@+id/today_fragment"
-        android:name="com.jeanbarrossilva.h2o.ui.today.TodayFragment"
-        android:label="@string/today" />
-</navigation>
